+++ aggregation-model/src/main/java/com/quicken/aggregation_model/vo/Summary/SummaryAbstract.java
@@ -1,10 +1,12 @@
 package com.quicken.aggregation_model.vo.Summary;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 
 @Getter
 public abstract class SummaryAbstract{
-    double income;
-    double expenses;
-    double net;
+    BigDecimal income;
+    BigDecimal expenses;
+    BigDecimal net;
 }
